requirements.txt
streamlit
python-dotenv
langchain
langgraph
langchain_groq
langchain_community
langchain_core
faiss-cpu
sentence-transformers
langchain-huggingface==0.0.3
pypdf
duckduckgo-search
beautifulsoup4
requests
tqdm
pytest
pydantic
numexpr
